pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bithub"
version = "1.0.0"
description = "Neural Net Link - Bithub Client System"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    "python-dotenv",
    "pytest"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.scripts]
bithub = "bithub.bithub:main"
